Split the issue URL budget helpers out and restore the console spy
